+++ jetson_nano/scripts/save_image.py
