+++ jetson_nano/scripts/test_camera.py
